src/backend_django/requirements.txt
# Django
Django==3.2.23
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
mysqlclient==2.2.1

# CORS
django-cors-headers==4.3.1

# MQTT
paho-mqtt==1.6.1

# Environment variables
python-decouple==3.8

# Timezone
pytz==2023.3

# Production server
gunicorn==21.2.0
whitenoise==6.6.0  # Para servir archivos estáticos en producción

# Development tools
django-debug-toolbar==4.2.0  # Solo en desarrollo

# Para trabajar con WebSockets (opcional para MQTT en tiempo real)
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0